offline -> online update
added new basic logic to checking to see if someone is coming back online and some more graceful exiting of the prgoram.

the biggest issue right now is that i'm not too sure if my port exiting is graceful enough, but i'm seemingly not able to reconnect to the same port so it makes it difficult to be able to test the offline/online capibilities.