offline messaging
ADDED:

Send_pending_message in server logic to be able to send pending messages (messages that have not been read)

mark as read logic so that for every message that i'm able to send once the user has come back online, it gets marked as read.

IMPROVEMENTS:

techincally it should be getting marked as read once EVERYBODY in the offline peers list gets the message, but this is good enough for the sake of this project. there are also just other things that can be a little more robust and clean up logic here and there, but wow. It's crazy how fun coding is when you're not just chatting